WIP: alternate meta scoring
This is implemented entirely in the standings page and is not reflected in
the solve map (which should be refactored to use the Standings code directly
instead of having its own implementation).

Note that this also fixes a bug that sorting the standings by other keys
other than rank was unstable.